# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet API",
        "description": "Example info",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://buckhill-test.test/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/admin/user-listing": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Get users",
                "description": "Get all non admin users",
                "operationId": "usersListing",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "true or false",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "search parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "column to sort with, e.g id, first_name, last_name, created_at",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All users fetched"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/user-edit/{uuid}": {
            "put": {
                "tags": [
                    "admin"
                ],
                "summary": "Update user record",
                "description": "Update user record",
                "operationId": "userUpdate",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "uuid of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input User Details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "address",
                                    "phone_number"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "adewale"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "charles"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "no 3, york lane"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "2348253796851"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "is_marketing": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/user-delete/{uuid}": {
            "delete": {
                "tags": [
                    "admin"
                ],
                "summary": "Delete user",
                "description": "Delete user record ",
                "operationId": "userDelete",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "uuid of product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Deleted"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}